feat: Add support for retrieving announcements and version from Notion database
The code changes in this commit add support for retrieving announcements and version information from a Notion database. The `fetch` function now checks the `action` parameter in the URL to determine whether to retrieve the version or announcements. If the action is "get_version", it retrieves the version information from the database. If the action is "get_announcements", it retrieves the announcements that are marked as published.

This commit also updates the project name and dependencies, as indicated in the recent user commits.